Add test utils class to avoid redundancy
