+++ tests/test_navigation.py
@@ -4,11 +4,12 @@ from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
+from test_utils import TestUtil
 import requests
 import os
 import time
 
-class TestNavigation(unittest.TestCase):
+class TestNavigation(unittest.TestCase, TestUtil):
 
     def setUp(self):
         """Set up testing"""
@@ -16,11 +17,14 @@ class TestNavigation(unittest.TestCase):
         self.driver.get("http://localhost:8000")
         self.email = os.getenv("TEST_EMAIL")
         self.password = os.getenv("TEST_PASSWORD")
+        self.wait = WebDriverWait(self.driver, 5)
         self.driver.refresh()
     
     def test_dashboard(self):
+
         """Test if dashboard contents are correct"""
         driver = self.driver
+        self.login(driver)
         assert driver.title == "Tech Radar Data Collection"
 
         time.sleep(1)
@@ -28,6 +32,7 @@ class TestNavigation(unittest.TestCase):
     def test_project_nav(self):
         """Test if project navigation is correct"""
         driver = self.driver
+        self.login(driver)
         driver.find_element(By.ID, "tab_add-projects").click()
         assert driver.current_url == "http://localhost:8000/dashboard#add-projects"
