+++ tests/test_project_creation.py
@@ -4,12 +4,13 @@ from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
+from test_utils import TestUtil
 import requests
 import os
 import random
 import time
 
-class TestProjectCreation(unittest.TestCase):
+class TestProjectCreation(unittest.TestCase, TestUtil):
 
     def setUp(self):
         """Set up testing"""
@@ -24,18 +25,7 @@ class TestProjectCreation(unittest.TestCase):
         """Test creating a project from beginning to end"""
         driver = self.driver
 
-        self.click_link(driver, "Sign in with Cognito")
-
-        form = self.wait.until(
-            EC.visibility_of_element_located((By.XPATH, "//div[contains(@class,'visible-md')]//form[@name='cognitoSignInForm']"))
-        )
-
-        email = form.find_element(By.NAME, "username")
-        password = form.find_element(By.NAME, "password")
-        email.send_keys(self.email)
-        password.send_keys(self.password)
-        button = form.find_element(By.NAME, "signInSubmitButton")
-        self.wait.until(EC.element_to_be_clickable(button)).click()
+        self.login(driver)
 
         driver.get("http://localhost:8000/survey")
 
@@ -87,20 +77,6 @@ class TestProjectCreation(unittest.TestCase):
 
     def tearDown(self):
         self.driver.close()
-    
-    def click_link(self, driver, link_text):
-        """Click link via text"""
-        link = driver.find_element(By.LINK_TEXT, link_text)
-        driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
-        self.wait.until(EC.element_to_be_clickable(link)).click()
-        driver.implicitly_wait(10)
-
-    def click_radio(self, driver, options):
-        """Randomly choose option from radio"""
-        choice = random.choice(options)
-        radio = driver.find_element(By.ID, choice)
-        self.wait.until(EC.element_to_be_clickable(radio)).click()
-        return choice
 
     def assert_validation_page(self):
         """Assert that the validation page has the correct details"""
